requirements.txt
eel==0.16.0
googletrans==3.0.0
hugchat==0.4.9
playsound==1.3.0
pvporcupine==3.0.2
pyaudio==0.2.14
pyautogui==0.9.54
pyttsx3==2.90
pywhatkit==5.4
speedtest==0.0.1
g4f==0.3.2.2
python_pptx==0.6.23
pyttsx3==2.90
SpeechRecognition==3.10.4
